+++ packages/studio-graph/src/components/CurvatureLinks/index.tsx
@@ -8,16 +8,22 @@ interface ILinkCurvatureProps {}
 
 const CurvatureLinks: React.FunctionComponent<ILinkCurvatureProps> = props => {
   const { store } = useContext();
-  const { graph, source, data } = store;
+  const { graph, data } = store;
   const [curvatured, setToogle] = React.useState(false);
   const handleClick = () => {
     if (graph) {
+      const { nodes, edges } = data;
       if (!curvatured) {
-        const { nodes, edges } = processLinks(data);
-        graph.linkCurvature('_style_curvature');
-        graph.graphData({ links: edges, nodes });
+        const links = processLinks(edges);
+        graph.linkCurvature('__style_curvature');
+        graph.graphData({ links, nodes });
       } else {
-        graph.graphData(Utils.fakeSnapshot({ nodes: source.nodes, links: source.edges }));
+        const links = edges.map(item => {
+          //@ts-ignore
+          const { __controlPoints, __style_curvature, __nodePairId, __indexColor, ...others } = item;
+          return others;
+        });
+        graph.graphData({ nodes, links });
       }
     }
     setToogle(!curvatured);
@@ -32,49 +38,54 @@ const CurvatureLinks: React.FunctionComponent<ILinkCurvatureProps> = props => {
 
 export default CurvatureLinks;
 
-function processLinks(graphData) {
-  const data = Utils.fakeSnapshot(graphData);
+function processLinks(edges) {
   let selfLoopLinks = {};
   let sameNodesLinks = {};
   const curvatureMinMax = 0.3; //0.5;
 
-  // 1. assign each link a nodePairId that combines their source and target independent of the edges direction
-  // 2. group edges together that share the same two nodes or are self-loops
-  data.edges.forEach(link => {
-    link.nodePairId = link.source <= link.target ? link.source + '_' + link.target : link.target + '_' + link.source;
-    let map = link.source === link.target ? selfLoopLinks : sameNodesLinks;
-    if (!map[link.nodePairId]) {
-      map[link.nodePairId] = [];
+  edges.forEach(link => {
+    const source = typeof link.source === 'object' ? link.source.id : link.source;
+    const target = typeof link.target === 'object' ? link.target.id : link.target;
+
+    const __nodePairId = `${source}_${target}`;
+    link.__nodePairId = __nodePairId;
+    delete link.__controlPoints;
+
+    let map = source === target ? selfLoopLinks : sameNodesLinks;
+
+    if (!map[__nodePairId]) {
+      map[__nodePairId] = [];
     }
-    map[link.nodePairId].push(link);
+
+    map[__nodePairId].push(link);
   });
 
-  // Compute the _style_curvature for self-loop edges to avoid overlaps
+  // Compute the __style_curvature for self-loop edges to avoid overlaps
   Object.keys(selfLoopLinks).forEach(id => {
     let edges = selfLoopLinks[id];
     let lastIndex = edges.length - 1;
-    edges[lastIndex]._style_curvature = 1;
+    edges[lastIndex].__style_curvature = 1;
     let delta = (1 - curvatureMinMax) / lastIndex;
     for (let i = 0; i < lastIndex; i++) {
-      edges[i]._style_curvature = curvatureMinMax + i * delta;
+      edges[i].__style_curvature = curvatureMinMax + i * delta;
     }
   });
 
-  // Compute the _style_curvature for edges sharing the same two nodes to avoid overlaps
+  // Compute the __style_curvature for edges sharing the same two nodes to avoid overlaps
   Object.keys(sameNodesLinks)
-    .filter(nodePairId => sameNodesLinks[nodePairId].length > 1)
-    .forEach(nodePairId => {
-      let edges = sameNodesLinks[nodePairId];
+    .filter(__nodePairId => sameNodesLinks[__nodePairId].length > 1)
+    .forEach(__nodePairId => {
+      let edges = sameNodesLinks[__nodePairId];
       let lastIndex = edges.length - 1;
       let lastLink = edges[lastIndex];
-      lastLink._style_curvature = curvatureMinMax;
+      lastLink.__style_curvature = curvatureMinMax;
       let delta = (2 * curvatureMinMax) / lastIndex;
       for (let i = 0; i < lastIndex; i++) {
-        edges[i]._style_curvature = -curvatureMinMax + i * delta;
+        edges[i].__style_curvature = -curvatureMinMax + i * delta;
         if (lastLink.source !== edges[i].source) {
-          edges[i]._style_curvature *= -1; // flip it around, otherwise they overlap
+          edges[i].__style_curvature *= -1; // flip it around, otherwise they overlap
         }
       }
     });
-  return data;
+  return edges;
 }