+++ packages/studio-importor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @graphscope/studio-importor
 
+## 0.1.12
+
+### Patch Changes
+
+- update studio graph
+- Updated dependencies
+  - @graphscope/studio-components@0.1.12
+  - @graphscope/studio-server@1.0.9
+  - @graphscope/use-zustand@0.1.4
+
 ## 0.1.11
 
 ### Patch Changes