+++ packages/studio-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphscope/studio-server
 
+## 1.0.9
+
+### Patch Changes
+
+- update studio graph
+
 ## 1.0.8
 
 ### Patch Changes