+++ packages/studio-website/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @graphscope/studio-site
 
+## 0.1.12
+
+### Patch Changes
+
+- update studio graph
+- Updated dependencies
+  - @graphscope/studio-importor@0.1.12
+  - @graphscope/studio-explore@0.1.10
+  - @graphscope/studio-graph@0.1.12
+  - @graphscope/graphy-website@0.1.6
+  - @graphscope/studio-components@0.1.12
+  - @graphscope/studio-driver@0.1.10
+  - @graphscope/studio-query@0.1.12
+  - @graphscope/studio-server@1.0.9
+  - @graphscope/use-zustand@0.1.4
+
 ## 0.1.11
 
 ### Patch Changes